% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_cluster.R
\name{contar_cluster}
\alias{contar_cluster}
\title{Conta quantos estão em cada cluster}
\usage{
contar_cluster(ano_num, tipo_matriz, nome_matriz)
}
\arguments{
\item{ano_num}{\code{numric}. Ano.}

\item{tipo_matriz}{Um objeto \code{listw} ou \code{matrix}. Uma matriz de pesos espaciais para o Espirito Santo.}

\item{nome_matriz}{\code{character}. Um rótulo que identifique a matriz de pesos espaciais.}
}
\value{
Um \code{tibble} que pode conter de 2 a 5 linhas e 4 colunas.
}
\description{
Conta quantos estão em cada cluster
}
\examples{
# Quatidade de municípios por cluster para o ano de 2015, usando a matriz IDW.
contar_cluster(2015, mat_idw, "IDW")

}
