% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_pvalor_teste_ML_robusto.R
\name{extrair_pvalor_teste_ML_robusto}
\alias{extrair_pvalor_teste_ML_robusto}
\title{Extrai o p-valor do teste ML robusto}
\usage{
extrair_pvalor_teste_ML_robusto(tipo_matriz, tipo_teste, nome_matriz)
}
\arguments{
\item{tipo_matriz}{\code{listw} ou \code{matrix}.  Uma matriz de pesos espaciais para o Espírito Santo.}

\item{tipo_teste}{"rlml" ou "rlme". Testar dependência espacial na variável dependente("rlml") ou no termo de erro ("rlme").}

\item{nome_matriz}{Um rótulo para a matriz de pesos espaciais.}
}
\value{
Um \code{tibble} com linha e 3 colunas.
}
\description{
Extrai o p-valor do teste ML robusto
}
\examples{
# p-valor do teste ML robusto para dependência espacial na variável dependente
extrair_pvalor_teste_ML_robusto(mat_idw, "rlml", "IDW")

}
