% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identificar_outliers.R
\name{identificar_outliers}
\alias{identificar_outliers}
\title{Identifica quais observações são outliers}
\usage{
identificar_outliers(ano, tipo_matriz, nome_matriz)
}
\arguments{
\item{ano}{\code{numeric}. Ano.}

\item{tipo_matriz}{\code{listw} ou \code{matrix}. Uma matriz de pesos espaciais para o Espirito Santo.}

\item{nome_matriz}{Um rótulo que identifique a matriz de pesos espaciais.}
}
\value{
Um \code{tibble} com 4 linhas e 4 colunas, contendo os outliers e a distância de Cook o ano selecionado.
}
\description{
Identifica quais observações são outliers
}
\examples{
# Verificando outliers e a distância de Cook para 2015.
identificar_outliers(2015, mat_idw, "IDW")

}
