% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_idw.R
\docType{data}
\name{mat_idw}
\alias{mat_idw}
\title{Matriz de pesos espaciais do tipo inverso da distância - municípios capixabas}
\format{
An object of class \code{listw} (inherits from \code{nb}) of length 4.
}
\usage{
mat_idw
}
\description{
Matriz de pesos espaciais do tipo inverso da distância dos 78 municípios capixabas.
}
\keyword{datasets}
