% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_idw_50_km.R
\docType{data}
\name{mat_idw_50_km}
\alias{mat_idw_50_km}
\title{Matriz de pesos espaciais do tipo inverso da distância considerando municípios em até 50 Km de distâncias - municípios capixabas}
\format{
An object of class \code{listw} (inherits from \code{nb}) of length 4.
}
\usage{
mat_idw_50_km
}
\description{
Matriz de pesos espaciais do tipo inverso da distância considerando municípios em até 50 Km de distâncias dos 78 municípios capixabas.
}
\keyword{datasets}
