% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_queen.R
\docType{data}
\name{mat_queen}
\alias{mat_queen}
\title{Matriz de pesos espaciais do tipo Queen - municípios capixabas}
\format{
An object of class \code{listw} (inherits from \code{nb}) of length 3.
}
\usage{
mat_queen
}
\description{
Matriz de pesos espaciais do tipo Queen dos 78 municípios capixabas.
}
\keyword{datasets}
