% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graf_disp.R
\name{plot_graf_disp}
\alias{plot_graf_disp}
\title{Plota o gráfico de dispersão de Moran}
\usage{
plot_graf_disp(dados)
}
\arguments{
\item{dados}{Um objeto \code{tibble} com 78 linhas e 43 colunas gerado pela função \link{prep_dados_aede}.}
}
\value{
Um gráfico de dispersão de Moran.
}
\description{
Plota o gráfico de dispersão de Moran
}
\examples{
# Plota o gráfico de dispersão de Moran
prep_dados_aede(2015, mat_queen) |> plot_graf_disp()

}
