% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapa.R
\name{plot_mapa}
\alias{plot_mapa}
\title{Plota o mapa de significância LISA}
\usage{
plot_mapa(dados, col)
}
\arguments{
\item{dados}{Um objeto \code{tibble} com 78 linhas e 43 colunas gerado pela função \link{prep_dados_aede}.}

\item{col}{Preencher com a variável fator \code{quadrante} do \code{tibble} retornado pela função \link{prep_dados_aede}.}
}
\value{
Um mapa de significância LISA do Espírito Santo.
}
\description{
Plota o mapa de significância LISA
}
\examples{
# Plotando o mapa de significância LISA.
prep_dados_aede(2015, mat_idw_100_km) |>
 plot_mapa(col = factor(quadrante))

}
