% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dados_aede.R
\name{prep_dados_aede}
\alias{prep_dados_aede}
\title{Prepara os dados para a AEDE}
\usage{
prep_dados_aede(year, tipo_matriz)
}
\arguments{
\item{year}{Seleciona o ano.}

\item{tipo_matriz}{Objeto do tipo \code{listw} ou \code{matrix} representado os pesos espaciais para os municípios do Espírito Santo. Usar os matrizes \code{mat_idw}, \code{mat_idw_100_km}, \code{mat_idw_50_km}, \code{mat_queen}i nclusas no pacote ou as funções  \code{spdep::nb2mat}, \code{spdep::nb2listw} ou \code{spdep::nb2listwdist} para criar uma matriz de pesos espaciais para o Espírito Santo.}
}
\value{
Um \code{data.frame} com 78 linhas e 43 colunas, contendo variáveis fiscais, demográficas, políticas, demográficas para os 78 municípios capixabas.
}
\description{
A função filtra o painel de dados por ano, além de calcular e agregar ao data.frame o I de Moran local, usado para classificar as observações em clusters do mapa de significância LISA e quadrantes do gráfico de dispersão de Moran.
}
\examples{
#Preparar os dados para o ano 2015 e usar a matriz `mat_idw`.
prep_dados_aede(2015, mat_idw)
}
