% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultados_teste_ML_robusto.R
\name{resultados_teste_ML_robusto}
\alias{resultados_teste_ML_robusto}
\title{Resultados dos testes ML robusto para dependência espacial no termo de erro e na variável dependente para todas as matrizes diponíveis.}
\usage{
resultados_teste_ML_robusto()
}
\value{
Um \code{tibble} com 4 linhas e 3 colunas.
}
\description{
Resultados dos testes ML robusto para dependência espacial no termo de erro e na variável dependente para todas as matrizes diponíveis.
}
\examples{
# Comparando os reusltados dos testes ML Erro Robusto e ML Lag Robusto
resultados_teste_ML_robusto()

}
