% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_munES_obj.R
\docType{data}
\name{sf_munES}
\alias{sf_munES}
\title{Dados Espaciais - Espírito Santo}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 78 rows and 2 columns.
}
\usage{
sf_munES
}
\description{
Simple feature collection with 78 features and 1 field
}
\keyword{datasets}
