% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todos_outliers.R
\name{todos_outliers}
\alias{todos_outliers}
\title{Identifica os outliers e a distância de Cook para todos os anos do painel e todas as opções de matriz de pesos espaciais disponíveis no pacote}
\usage{
todos_outliers()
}
\value{
Uma tabela com os outliers e a distância de Cook.
}
\description{
Identifica os outliers e a distância de Cook para todos os anos do painel e todas as opções de matriz de pesos espaciais disponíveis no pacote
}
\examples{
# Identificando outliers e distância de Cook para todos os anos e matrizes
todos_outliers()

}
